adjust styled-components
