+++ front/README.md
@@ -1,11 +1,5 @@
 # Project Cainthus
 
-Ps: Sorry about my unique commit, I was read the README.md about the test and I jumped the part about the branch and started to write the code and to commit in the master. I was doing my final checkout and I saw that I was doing wrong, I did a revert and create the new branch but I lost my commits. Sorry again. 
-
-## Challenge
-
-[Readme with description the challenge](/README-TEST.md)
-
 ## Technologies to front
 
 - React: Lib single application
@@ -15,17 +9,17 @@ Ps: Sorry about my unique commit, I was read the README.md about the test and I 
 - Webpack 4: A bundler for javascript
 - Babel: Javascript compiler
 - Jest: Test all JavaScript code
-- Chai: Assertion library (better than jest)
-- Sinon: Standalone test spies, stubs and mocks for JavaScript
 - React-thunk: Async middleware for Redux.
-- MomentJs: Parse, validate, manipulate, and display dates and times in JavaScript.
 - Enzyme: Enzyme is a JavaScript Testing utility for React that makes it easier to assert, manipulate, and traverse your React Components' output.
+- Storybook: UI component dev & test: React, React Native, Vue, Angular, Ember & more!
 
 I didn't use react-lazyload because this project is very small, but if it is necessary to the project is very easy to implement after.
 
-## Run Project
+## Run Front-end Project
 
   To run this project you need to install all packages in npm, to do it run:
+
+`$ npm install`
 
 `$ npm install`
 
@@ -57,12 +51,4 @@ Test with watch (dev-mode)
 
 `$ npm run test:watch`
 
-## Deploy
-
-To do the deploy, you need only install and do your login in the Now.
-
-`$ npm install -g now`
-
-`$ now login`
-
-`$ now`
+### Storybook
