+++ front/src/components/card/components/card.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components'
+
+const Card = styled.div`
+  position: relative;
+  width: 100%;
+  height: calc(100vh - 7rem);
+  background: #fff;
+  border-radius: .4rem;
+  min-width: 27.5rem;
+  box-sizing: border-box;
+  padding-bottom: 2.5rem;
+`
+
+export default Card
