+++ front/src/components/circle-button/components/logo.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+import Logo from '../plus.svg'
+
+const LogoStyle = styled(Logo)`
+  fill: #fff;
+  & path {
+   fill: #fff;
+  }
+`
+
+export default LogoStyle
