+++ front/src/components/empty/components/description.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const Descriptions = styled.p`
+  font-size: 1.5rem;
+  color: #707070;
+  text-align: center;
+`
+
+export default Descriptions
