+++ front/src/components/empty/components/empty.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+const Empty = styled.div`
+  width: 100%;
+  height: calc(100% - 2rem);
+  box-sizing: border-box;
+  padding: 1.5rem;
+  overflow: auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+`
+
+export default Empty
