+++ front/src/components/error/components/description.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+const Descriptions = styled.p`
+  font-size: 1.5rem;
+  color: #707070;
+`
+
+export default Descriptions
