+++ front/src/components/grid-item/components/description.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const Description = styled.p`
+  color: #707070;
+  font-size: 1.2rem;
+  margin: 0 0 1rem;
+`
+
+export default Description
