+++ front/src/components/grid-item/components/info-container.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const InfosContainer = styled.div`
+  display: flex;
+  align-items: center;
+  padding: 0 1rem;
+`
+
+export default InfosContainer
