+++ front/src/components/grid-item/components/title-container.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const TitlesContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-left: .5rem;
+`
+
+export default TitlesContainer
