+++ front/src/components/grid-item/components/title.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+const Title = styled.h1`
+  color: #707070;
+  font-size: 1.6rem;
+  font-weight: 400;
+  display: inline-block;
+  margin-top: .7rem;
+  margin-bottom: .5rem;
+  vertical-align: top;
+`
+
+export default Title
