+++ front/src/components/grid/components/grid.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const Grid = styled.ul`
+  list-style: none;
+  width: 100%;
+  height: calc(100% - 2rem);
+  box-sizing: border-box;
+  padding: 1.5rem;
+  overflow: auto;
+`
+
+export default Grid
