+++ front/src/components/input/components/input-container.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const InputContainer = styled.div`
+  width: 100%;
+  position: relative;
+  margin: 1rem;
+  box-sizing: border-box;
+`
+
+export default InputContainer
