+++ front/src/components/input/components/label.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const Label = styled.label`
+  font-family: "Roboto", "Helvetica", "Arial", sans-serif;
+  font-size: 1.2rem;
+  color: rgba(0, 0, 0, 0.54);
+`
+
+export default Label
