+++ front/src/components/loading/components/container-spinner.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+const ContainerSpinner = styled.div`
+  width: 60px;
+  height: 15px;
+  position: relative;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+`
+
+export default ContainerSpinner
