+++ front/src/components/loading/components/modal-container.js
@@ -1,0 +1,13 @@
+import styled from 'styled-components'
+
+const ModalContainer = styled.div`
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, .5);
+  z-index: 200;
+`
+
+export default ModalContainer
