+++ front/src/components/modal/components/animation.js
@@ -1,0 +1,13 @@
+import {keyframes} from 'styled-components'
+
+const Animation = keyframes`
+  0% {
+    opacity: 0;
+  }
+
+  100% {
+    opacity: 1;
+  }
+`
+
+export default Animation
