+++ front/src/components/title/index.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Title = styled.h1`
+  font-size: 2rem;
+  color: #707070;
+  margin: 1.5rem;
+  font-weight: 600;
+`
+
+export default Title
