+++ front/src/containers/todo/components/button-form.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+import ButtonComponent from '../../../components/button'
+
+const ButtonForm = styled(ButtonComponent)`
+  left: 50%;
+  position: relative;
+  transform: translateX(-50%);
+`
+
+export default ButtonForm
