+++ front/src/containers/todo/components/button.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+
+const Button = styled.div`
+  position: absolute;
+  bottom: -2.5rem;
+  left: 50%;
+  transform: translateX(-50%);
+`
+
+export default Button
