+++ front/src/containers/todo/components/inputs-container.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+
+const InputsContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  margin-top: 1rem;
+`
+
+export default InputsContainer
