+++ front/src/containers/todo/components/modal-container.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+const ModalContainer = styled.div`
+  width: 35vw;
+
+  @media only screen and (max-width: 768px) {
+    width: 90vw;
+  }
+`
+
+export default ModalContainer
