+++ front/src/containers/todo/components/title.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+const Title = styled.h1`
+  padding-top: 1.5rem;
+  font-size: 2rem;
+  color: #707070;
+  text-align: center;
+  font-weight: 600;
+`
+
+export default Title
